requirements.txt
Flask
Flask-Mail
flask-compress
flask_assets
pymongo
scikit-learn
numpy
Pillow
ocrmypdf
tesseract
